requirements.txt
aiogrpc==1.6
alabaster==0.7.12
arrow==0.14.7
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
commonmark==0.9.0
coverage==4.5.4
croniter==0.3.30
cryptography==2.7
datadispatch==1.0.0
-e git+https://github.com/HarryThomas17/A-Decentralised-Digital-Identity-Architecture.git#egg=Decentralised_DigID
docutils==0.15.2
fabric-sdk-py==0.8.1
filelock==3.0.12
Flask==1.1.1
Flask-JWT-Extended==3.23.0
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
future==0.17.1
grpcio==1.24.0
hkdf==0.0.3
hypothesis==4.36.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
lark-parser==0.7.1
MarkupSafe==1.1.1
more-itertools==7.2.0
mpmath==1.1.0
msgpack-python==0.5.6
numpy==1.17.2
packaging==19.1
pluggy==0.12.0
protobuf==3.9.2
py==1.8.0
pycparser==2.19
pycryptodome==3.9.0
pycryptodomex==3.9.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pysha3==1.0b1
pytest==5.1.1
pytest-cov==2.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-nginx==1.5.3
pytz==2019.2
PyYAML==5.1.2
recommonmark==0.6.0
requests==2.22.0
Rx==3.0.1
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.8
toml==0.10.0
tox==3.14.0
urllib3==1.25.3
uWSGI==2.0.18
virtualenv==16.7.5
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0